Fix bug when export file
When exporting a file to csv a rather obtuse method was used utilizing
numpy save text feature, first a numpy array was made to store the
string values of every cell and then the np.savetxt method was called.
This can result in unnacurate info when commas are important as a part of
the cell value because by default commas were stripped out. The
following was done to fix the issue:

◆ Use context manager to open a new csv file in write mode
◆ Instantiate a csv.writer object
◆ Use csv.writer writerow method to write each row of data to file
◆ Replace the numpy method with above method