Update MyWidgets to conform to new model
Calls to the setData method need to use the new signature. Develop aux
methods namely prepareFormulas and rebuildFormulas that deal with
weakrefs objs so there are no weakrefs while pickling since it raises an
AttributeError. This methods will substitute weakrefs for appropiate
objs and later reconstruct those weakrefs when loading.