Optimize calculate
When recalculating a formula via pressing enter on the evaluate bar the
algorithm recreated the formula even though it didn't change just to
discard it at the end once it checked that it was the same. A check is
now done before to ensure that only a formula is recreated if it differs in its
text definition.